Update config_schema.py
changed step from int to float

+++ src/config_schema.py
@@ -95,7 +95,7 @@ class DynamicPromptSettings:
 # -------------------
 @dataclass
 class GenPlanStage:
-    step: int = 0
+    step: float = 0
     run_dir: str | None = None
     subtask: str = "actual_gen"
     DEBUG: bool = False
